Refactor RHF energy
