+++ src/rhf/Energy/EnergyHelpers.jl
