+++ src/rhf/JCRHF.jl
@@ -7,60 +7,7 @@ wave function for closed-shell systems.
 """
 module JCRHF
 
-using JuliaChem.JCModules
-using JuliaChem.JERI
-
-using MPI
-using JSON
-
-Base.include(@__MODULE__,"RHFHelpers.jl")
-Base.include(@__MODULE__,"RHFSCF.jl")
-
-"""
-  run(input_info::Dict{String,Dict{String,Any}}, basis::Basis)
-
-Execute the JuliaChem RHF algorithm.
-
-One input variable is required:
-1. input_info = Information gathered from the input file.
-2. basis = The basis set shells, determined from the input file.
-
-One variable is output:
-1. scf = Data saved from the SCF calculation.
-
-Thus, proper use of the RHF.run() function would look like this:
-
-```
-scf = RHF.run(input_info, basis)
-```
-"""
-function run(mol::Molecule, basis::Basis, 
-  scf_flags; output="none")
-  
-  comm=MPI.COMM_WORLD
-
-  if MPI.Comm_rank(comm) == 0 && output == "verbose"
-    println("--------------------------------------------------------------------------------")
-    println("                       ========================================                 ")
-    println("                                RESTRICTED CLOSED-SHELL                         ")
-    println("                                  HARTREE-FOCK ENERGY                           ")
-    println("                       ========================================                 ")
-    println("")
-  end
-
-  #== actually perform scf calculation ==#
-  rhfenergy = rhf_energy(mol, basis, scf_flags; output=output)
-
-  if MPI.Comm_rank(comm) == 0 && output == "verbose"
-    println("                       ========================================                 ")
-    println("                              END RESTRICTED CLOSED-SHELL                       ")
-    println("                                  HARTREE-FOCK ENERGY                           ")
-    println("                       ========================================                 ")
-    println("--------------------------------------------------------------------------------")
-  end
-
-  return rhfenergy 
-end
-export run
+Base.include(@__MODULE__,"Energy/Energy.jl")
+export Energy
 
 end
